requirements.txt
boto3
fastapi==0.115.0
ffmpeg-python==0.2.0
numpy==1.26.4
pyannote.audio==3.3.2
pydantic==2.8.2
python-multipart==0.0.9
torch<2.6
torchaudio
uvicorn[standard]==0.30.6
whisperx==3.1.5
